feat(domain): Places APIをドメイン層で利用するためのサービスのパッケージ名を変更
