+++ internal/domain/services/placesearch/search_nearby_location.go
@@ -1,4 +1,4 @@
-package place
+package placesearch
 
 import (
 	"context"
