+++ internal/domain/services/placesearch/searched_places.go
@@ -1,4 +1,4 @@
-package place
+package placesearch
 
 import (
 	"context"
