+++ internal/domain/services/plangen/service.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"go.uber.org/zap"
 	"poroto.app/poroto/planner/internal/domain/repository"
-	"poroto.app/poroto/planner/internal/domain/services/place"
+	"poroto.app/poroto/planner/internal/domain/services/placesearch"
 	"poroto.app/poroto/planner/internal/domain/utils"
 	"poroto.app/poroto/planner/internal/infrastructure/api/google/places"
 	"poroto.app/poroto/planner/internal/infrastructure/api/openai"
@@ -14,7 +14,7 @@ import (
 
 type Service struct {
 	placesApi                  places.PlacesApi
-	placeService               place.Service
+	placeService               placesearch.Service
 	planCandidateRepository    repository.PlanCandidateRepository
 	openaiChatCompletionClient openai.ChatCompletionClient
 	logger                     *zap.Logger
@@ -26,7 +26,7 @@ func NewService(ctx context.Context) (*Service, error) {
 		return nil, fmt.Errorf("error while initizalizing Places api: %v", err)
 	}
 
-	placeService, err := place.NewPlaceService(ctx)
+	placeService, err := placesearch.NewPlaceService(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("error while initializing place service: %v", err)
 	}
